+++ CHANGELOG.md
@@ -7,3 +7,9 @@
 * Now use the WgApiFactory
 * Response return JsonResponse
 * Add logging
+
+### [2.1.0]() (2020-11-14)
+* Add properties file
+* Add proxy settings
+* Refectories WgApiUriBuilder class
+